backend/vehicle-police-service/requirements.txt
fastapi[standard]
email_validator>=2.2.0
gunicorn>=23.0.0
sqlalchemy>=2.0.41
psycopg2>=2.9.10
alembic>=1.16.4
python-jose>=3.5.0
structlog>=25.5.0
passlib[argon2]
typer>=0.16.0